+++ racio/.env.example
@@ -9,17 +9,6 @@ SECRET_KEY=sk-9f73s3ljTXVcMT3Blb3ljTqtsKiGHXVcMT3BlbkFJLK7U
 CONSOLE_API_URL=http://127.0.0.1:5002
 CONSOLE_WEB_URL=http://127.0.0.1:5002
 
-# Service API base URL
-SERVICE_API_URL=http://127.0.0.1:5002
-
-# Web APP base URL
-APP_WEB_URL=http://127.0.0.1:3000
-
-# Files URL
-#FILES_URL=http://127.0.0.1:5002
-
-# celery configuration
-#CELERY_BROKER_URL=
 
 # redis configuration
 REDIS_HOST=
